Ensure netlink messages' buffer has correct alignment before parsing.
Netlink requires that each message, and message part, start address is (4B) aligned.
However, there is no guarantee that the last message in a buffer is padded to the alignment.
Thus, the filled buffer may have fewer bytes than required by strict adherence to alignment.

The change ensures that messages passed for parsing are aligned as expected.
As the input buffer size (8192) is alrady correctly aligned, we are guaranteed
that the added alignment would not overflow the buffer.